+++ outputs.tf
@@ -1,1 +1,7 @@
+output "vpc_endpoint_sgs" {
+  value = aws_security_group.this
+}
 
+output "vpc_endpoint_interface_services" {
+  value = aws_vpc_endpoint.interface_services
+}